+++ spec/form_params_spec.rb
@@ -15,6 +15,7 @@ describe 'Form Params' do
       params do
         requires :id, type: Integer, desc: 'id of item'
         requires :name, type: String, desc: 'name of item'
+        requires :conditions, type: Integer, desc: 'conditions of item', values: [1,2,3]
       end
       put '/items/:id' do
         {}
@@ -23,6 +24,7 @@ describe 'Form Params' do
       params do
         requires :id, type: Integer, desc: 'id of item'
         requires :name, type: String, desc: 'name of item'
+        optional :conditions, type: String, desc: 'conditions of item', values: Proc.new {['1','2']}
       end
       patch '/items/:id' do
         {}
@@ -38,6 +40,7 @@ describe 'Form Params' do
   end
 
   it 'retrieves the documentation form params' do
+    puts subject['apis']
     expect(subject['apis']).to eq([
       {
         'path' => '/items.{format}',
@@ -59,7 +62,11 @@ describe 'Form Params' do
             'summary' => '',
             'nickname' => 'PUT-items--id---format-',
             'method' => 'PUT',
-            'parameters' => [{ 'paramType' => 'path', 'name' => 'id', 'description' => 'id of item', 'type' => 'integer', 'required' => true, 'allowMultiple' => false, 'format' => 'int32' }, { 'paramType' => 'form', 'name' => 'name', 'description' => 'name of item', 'type' => 'string', 'required' => true, 'allowMultiple' => false }],
+            'parameters' => [
+                { 'paramType' => 'path', 'name' => 'id', 'description' => 'id of item', 'type' => 'integer', 'required' => true, 'allowMultiple' => false, 'format' => 'int32' },
+                { 'paramType' => 'form', 'name' => 'name', 'description' => 'name of item', 'type' => 'string', 'required' => true, 'allowMultiple' => false },
+                { 'paramType' => 'form', 'name' => 'conditions', 'description' => 'conditions of item', 'type' => 'integer', 'required' => true, 'allowMultiple' => false, 'format' => 'int32', 'enum' => [1,2,3] }
+            ],
             'type' => 'void'
           },
           {
@@ -67,7 +74,11 @@ describe 'Form Params' do
             'summary' => '',
             'nickname' => 'PATCH-items--id---format-',
             'method' => 'PATCH',
-            'parameters' => [{ 'paramType' => 'path', 'name' => 'id', 'description' => 'id of item', 'type' => 'integer', 'required' => true, 'allowMultiple' => false, 'format' => 'int32' }, { 'paramType' => 'form', 'name' => 'name', 'description' => 'name of item', 'type' => 'string', 'required' => true, 'allowMultiple' => false }],
+            'parameters' => [
+                { 'paramType' => 'path', 'name' => 'id', 'description' => 'id of item', 'type' => 'integer', 'required' => true, 'allowMultiple' => false, 'format' => 'int32' },
+                { 'paramType' => 'form', 'name' => 'name', 'description' => 'name of item', 'type' => 'string', 'required' => true, 'allowMultiple' => false },
+                { 'paramType' => 'form', 'name' => 'conditions', 'description' => 'conditions of item', 'type' => 'string', 'required' => false, 'allowMultiple' => false, 'enum' => ['1','2'] }
+            ],
             'type' => 'void'
           }
         ]